+++ mods/BrewLAN/documentation/Changelog.md
@@ -272,6 +272,11 @@ All changes by Sean Wheeldon (Balthazar) unless otherwise stated.
   * Redesigned the Aeon energy storage.
   * Set the collision model of the Aeon energy storage to be spherical.
 
+  ### 2018-06-06
+
+  * Redesigned the Aeon mass storage.
+  * Set the collision model of the Aeon mass storage to be spherical.
+
 ## 2019-01-03|84 (FAF minor patch)
 
 * Created a function to add the FAF category CQUEMOV to function to mobile experimental mobile units.